Create range input and function for creating grid
This will create a range input as well as the script to take the value
and add it to a para so it is visible to the user. It will also add the
function to add a variable number of divs to the grid wrapper.

Currently trying to adjust the flex-basis dynamically to unsure that the
divs are square.