Create a grid that is dynamically changed using slider
This commit will replace a lot of existing code in both the styles and
script document. I came to the conclusion that rather then fussing with
flexbox why not use grid to create a grid. So that is what this commit
achieves by replacing code in both the styles and script files a grid is
created that is dynamically changeable by the user.